Consolidate keep-alive wrapper tests and trim comments
Move the with_keep_alive unit tests into test_sse_writer.py (the unit home for _sse.py) and drop the separate test file. Reword source and test comments to describe current behavior.
