+++ sdk/agentserver/azure-ai-agentserver-responses/tests/unit/test_sse_writer.py
@@ -4,7 +4,14 @@
 
 from __future__ import annotations
 
+import asyncio
+import contextvars
+from typing import AsyncIterator, List
+
 from azure.ai.agentserver.responses.streaming import _sse
+from azure.ai.agentserver.responses.streaming._sse import encode_keep_alive_comment, with_keep_alive
+
+_KA = encode_keep_alive_comment()
 
 
 class _FakeEvent:
@@ -59,3 +66,96 @@ def test_sse_writer__injects_monotonic_sequence_numbers() -> None:
 
     assert seq_first == 0, f"first sequence_number must be 0 for a new stream, got {seq_first}"
     assert seq_second == 1, f"second sequence_number must be 1 for a new stream, got {seq_second}"
+
+
+async def _collect(agen: AsyncIterator[str]) -> List[str]:
+    return [item async for item in agen]
+
+
+async def test_with_keep_alive__passthrough_when_disabled() -> None:
+    """A None or 0 interval yields the source unchanged with no keep-alive frames."""
+
+    async def source() -> AsyncIterator[str]:
+        yield "a"
+        yield "b"
+
+    assert await _collect(with_keep_alive(source(), None)) == ["a", "b"]
+    assert await _collect(with_keep_alive(source(), 0)) == ["a", "b"]
+
+
+async def test_with_keep_alive__emits_heartbeat_on_idle_gap() -> None:
+    """A keep-alive frame is emitted when the source stays idle past the interval."""
+
+    async def source() -> AsyncIterator[str]:
+        yield "a"
+        await asyncio.sleep(0.25)
+        yield "b"
+
+    out = await _collect(with_keep_alive(source(), 0.05))
+    assert _KA in out
+    assert [item for item in out if item != _KA] == ["a", "b"]
+    assert out.index("a") < out.index("b")
+
+
+async def test_with_keep_alive__preserves_item_order_across_gaps() -> None:
+    """Items arriving across several idle gaps keep their order and count."""
+
+    async def source() -> AsyncIterator[str]:
+        for i in range(5):
+            await asyncio.sleep(0.03)
+            yield f"item-{i}"
+
+    out = await _collect(with_keep_alive(source(), 0.01))
+    assert [item for item in out if item != _KA] == [f"item-{i}" for i in range(5)]
+
+
+async def test_with_keep_alive__preserves_contextvars_across_yields() -> None:
+    """A ContextVar the source sets once stays set across all of its yields, because the
+    source is advanced by a single task."""
+
+    cvar: contextvars.ContextVar[str] = contextvars.ContextVar("cvar", default="UNSET")
+    seen: List[str] = []
+
+    async def source() -> AsyncIterator[str]:
+        cvar.set("SET")
+        seen.append(cvar.get())
+        yield "a"
+        seen.append(cvar.get())
+        yield "b"
+        seen.append(cvar.get())
+
+    out = await _collect(with_keep_alive(source(), 5))
+    assert out == ["a", "b"]
+    assert seen == ["SET", "SET", "SET"]
+
+
+async def test_with_keep_alive__runs_source_finally_on_early_close() -> None:
+    """Closing the wrapper after one item runs the source's finally before returning."""
+
+    finalized = asyncio.Event()
+
+    async def source() -> AsyncIterator[str]:
+        try:
+            yield "a"
+            await asyncio.Event().wait()
+            yield "b"
+        finally:
+            finalized.set()
+
+    agen = with_keep_alive(source(), 5)
+    assert await agen.__anext__() == "a"
+    await agen.aclose()
+    await asyncio.wait_for(finalized.wait(), timeout=1.0)
+
+
+async def test_with_keep_alive__emits_heartbeat_while_source_idle_from_start() -> None:
+    """A source idle from the start still emits keep-alive frames, matching a reconnect
+    GET subscribed to an in-flight run that has not emitted since the cursor."""
+
+    async def idle_source() -> AsyncIterator[str]:
+        await asyncio.Event().wait()
+        yield "never"  # pragma: no cover
+
+    agen = with_keep_alive(idle_source(), 0.05)
+    assert await asyncio.wait_for(agen.__anext__(), timeout=1.0) == _KA
+    await agen.aclose()
